【提交摘要】v0.0.12 【详细内容】  - 新增：qerror包，快速格式化输出日常日志，并记录日志到文件  - 优化：框架启动后，自动处理panic异常并格式化输出日志  - 优化：移除mc包，并将原有的mc包移动到example中
