+++ example/bll/labItem/model.go
