+++ example/bll/laboratory/model.go
