+++ example/bll/liteLab/model.go
