+++ example/main.go
@@ -2,7 +2,8 @@ package main
 
 import (
 	"github.com/UritMedical/qf"
-	"github.com/UritMedical/qf/mc/patient"
+	"github.com/UritMedical/qf/example/bll/patient"
+	"github.com/UritMedical/qf/example/bll/sqlserver"
 )
 
 func main() {
@@ -12,5 +13,5 @@ func main() {
 func regBll(s *qf.Service) {
 	// 注册自定义业务
 	s.RegBll(&patient.Bll{}, "") // 患者信息业务
-	s.RegBll(&Bll{}, "")
+	s.RegBll(&sqlserver.Bll{}, "")
 }
